requirements.txt
deepface==0.0.79
keyboard==0.13.5
numpy==1.24.3
opencv_python==4.7.0.72
Pillow==9.5.0
Pillow==9.5.0
